move DigestData class into seperate file
