+++ src/Symfony/Component/Security/Http/Firewall/DigestAuthenticationListener.php
@@ -131,86 +131,3 @@ class DigestAuthenticationListener implements ListenerInterface
         $event->setResponse($this->authenticationEntryPoint->start($request, $authException));
     }
 }
-
-class DigestData
-{
-    private $elements;
-    private $header;
-    private $nonceExpiryTime;
-
-    public function __construct($header)
-    {
-        $this->header = $header;
-        $parts = preg_split('/, /', $header);
-        $this->elements = array();
-        foreach ($parts as $part) {
-            list($key, $value) = explode('=', $part);
-            $this->elements[$key] = '"' === $value[0] ? substr($value, 1, -1) : $value;
-        }
-    }
-
-    public function getResponse()
-    {
-        return $this->elements['response'];
-    }
-
-    public function getUsername()
-    {
-        return $this->elements['username'];
-    }
-
-    public function validateAndDecode($entryPointKey, $expectedRealm)
-    {
-        if ($keys = array_diff(array('username', 'realm', 'nonce', 'uri', 'response'), array_keys($this->elements))) {
-            throw new BadCredentialsException(sprintf('Missing mandatory digest value; received header "%s" (%s)', $this->header, implode(', ', $keys)));
-        }
-
-        if ('auth' === $this->elements['qop']) {
-            if (!isset($this->elements['nc']) || !isset($this->elements['cnonce'])) {
-                throw new BadCredentialsException(sprintf('Missing mandatory digest value; received header "%s"', $this->header));
-            }
-        }
-
-        if ($expectedRealm !== $this->elements['realm']) {
-            throw new BadCredentialsException(sprintf('Response realm name "%s" does not match system realm name of "%s".', $this->elements['realm'], $expectedRealm));
-        }
-
-        if (false === $nonceAsPlainText = base64_decode($this->elements['nonce'])) {
-            throw new BadCredentialsException(sprintf('Nonce is not encoded in Base64; received nonce "%s".', $this->elements['nonce']));
-        }
-
-        $nonceTokens = explode(':', $nonceAsPlainText);
-
-        if (2 !== count($nonceTokens)) {
-            throw new BadCredentialsException(sprintf('Nonce should have yielded two tokens but was "%s".', $nonceAsPlainText));
-        }
-
-        $this->nonceExpiryTime = $nonceTokens[0];
-
-        if (md5($this->nonceExpiryTime.':'.$entryPointKey) !== $nonceTokens[1]) {
-            new BadCredentialsException(sprintf('Nonce token compromised "%s".', $nonceAsPlainText));
-        }
-    }
-
-    public function calculateServerDigest($password, $httpMethod)
-    {
-        $a2Md5 = md5(strtoupper($httpMethod).':'.$this->elements['uri']);
-        $a1Md5 = md5($this->elements['username'].':'.$this->elements['realm'].':'.$password);
-
-        $digest = $a1Md5.':'.$this->elements['nonce'];
-        if (!isset($this->elements['qop'])) {
-        } elseif ('auth' === $this->elements['qop']) {
-            $digest .= ':'.$this->elements['nc'].':'.$this->elements['cnonce'].':'.$this->elements['qop'];
-        } else {
-            throw new \InvalidArgumentException('This method does not support a qop: "%s".', $this->elements['qop']);
-        }
-        $digest .= ':'.$a2Md5;
-
-        return md5($digest);
-    }
-
-    public function isNonceExpired()
-    {
-        return $this->nonceExpiryTime < microtime(true);
-    }
-}
